requirements.txt
boto3
pandas
apache-airflow-providers-apache-spark
pytest